refactor: create Swipeable composable
